+++ app/controllers/manage/pages_controller.rb
@@ -9,12 +9,12 @@ module Manage
     def flow
       @booking_states = current_organisation.booking_flow_class.state_classes.values
       @templates = @booking_states.index_with do |booking_state|
-        booking_state.templates.pluck(:key)
+        booking_state.rich_text_templates.keys
       end
       @templates[BookingStates::DefinitiveRequest] << :contract_text
       @templates[BookingStates::DefinitiveRequest] << :invoices_deposit_text
-      @templates[BookingStates::AwaitingContract] = [:awaiting_contract_notification]
-      @templates[BookingStates::Past] = %i[invoices_invoice_text payment_due_notification]
+      @templates[BookingStates::AwaitingContract] = [:email_contract_notification]
+      @templates[BookingStates::Past] = %i[invoices_invoice_text email_invoice_notification]
     end
   end
 end