+++ app/domain/booking_states/awaiting_tenant.rb
@@ -2,8 +2,8 @@
 
 module BookingStates
   class AwaitingTenant < Base
-    templates << MailTemplate.define(:awaiting_tenant_notification, context: %i[booking])
-    templates << MailTemplate.define(:booking_agent_request_accepted_notification, context: %i[booking])
+    use_mail_template(:awaiting_tenant_notification, context: %i[booking])
+    use_mail_template(:booking_agent_request_accepted_notification, context: %i[booking])
 
     def checklist
       []