+++ app/domain/booking_states/declined_request.rb
@@ -2,7 +2,7 @@
 
 module BookingStates
   class DeclinedRequest < Base
-    templates << MailTemplate.define(:declined_request_notification, context: %i[booking])
+    use_mail_template(:declined_request_notification, context: %i[booking])
 
     def checklist
       []