+++ app/models/invoice.rb
@@ -4,29 +4,31 @@
 #
 # Table name: invoices
 #
-#  id                   :bigint           not null, primary key
-#  amount               :decimal(, )      default(0.0)
-#  amount_open          :decimal(, )
-#  discarded_at         :datetime
-#  issued_at            :datetime
-#  locale               :string
-#  payable_until        :datetime
-#  payment_info_type    :string
-#  payment_ref          :string
-#  payment_required     :boolean          default(TRUE)
-#  ref                  :string
-#  sent_at              :datetime
-#  sequence_number      :integer
-#  sequence_year        :integer
-#  text                 :text
-#  type                 :string
-#  created_at           :datetime         not null
-#  updated_at           :datetime         not null
-#  booking_id           :uuid
-#  supersede_invoice_id :bigint
+#  id                        :bigint           not null, primary key
+#  amount                    :decimal(, )      default(0.0)
+#  amount_open               :decimal(, )
+#  discarded_at              :datetime
+#  issued_at                 :datetime
+#  locale                    :string
+#  payable_until             :datetime
+#  payment_info_type         :string
+#  payment_ref               :string
+#  payment_required          :boolean          default(TRUE)
+#  ref                       :string
+#  sent_at                   :datetime
+#  sequence_number           :integer
+#  sequence_year             :integer
+#  text                      :text
+#  type                      :string
+#  created_at                :datetime         not null
+#  updated_at                :datetime         not null
+#  booking_id                :uuid
+#  sent_with_notification_id :bigint
+#  supersede_invoice_id      :bigint
 #
 
 class Invoice < ApplicationRecord
+  extend RichTextTemplate::Definition
   include Subtypeable
   include Discard::Model
 
@@ -35,6 +37,7 @@ class Invoice < ApplicationRecord
 
   belongs_to :booking, inverse_of: :invoices, touch: true
   belongs_to :supersede_invoice, class_name: :Invoice, optional: true, inverse_of: :superseded_by_invoices
+  belongs_to :sent_with_notification, class_name: 'Notification', optional: true
 
   has_many :invoice_parts, -> { ordered }, inverse_of: :invoice, dependent: :destroy
   has_many :superseded_by_invoices, class_name: :Invoice, dependent: :nullify,
@@ -52,7 +55,7 @@ class Invoice < ApplicationRecord
   scope :unsettled, -> { kept.where.not(type: 'Invoices::Offer').where.not(arel_table[:amount_open].eq(0)) }
   scope :refund,    -> { kept.where(arel_table[:amount_open].lt(0)) }
   scope :paid,      -> { kept.where(arel_table[:amount_open].lteq(0)) }
-  scope :sent,      -> { where.not(sent_at: nil) }
+  scope :sent, -> { where.not(sent_at: nil) }
   scope :unsent,    -> { kept.where(sent_at: nil) }
   scope :overdue,   ->(at = Time.zone.today) { kept.where(arel_table[:payable_until].lteq(at)) }
   scope :of,        ->(booking) { where(booking:) }