+++ app/models/invoices/deposit.rb
@@ -4,26 +4,27 @@
 #
 # Table name: invoices
 #
-#  id                   :bigint           not null, primary key
-#  amount               :decimal(, )      default(0.0)
-#  amount_open          :decimal(, )
-#  discarded_at         :datetime
-#  issued_at            :datetime
-#  locale               :string
-#  payable_until        :datetime
-#  payment_info_type    :string
-#  payment_ref          :string
-#  payment_required     :boolean          default(TRUE)
-#  ref                  :string
-#  sent_at              :datetime
-#  sequence_number      :integer
-#  sequence_year        :integer
-#  text                 :text
-#  type                 :string
-#  created_at           :datetime         not null
-#  updated_at           :datetime         not null
-#  booking_id           :uuid
-#  supersede_invoice_id :bigint
+#  id                        :bigint           not null, primary key
+#  amount                    :decimal(, )      default(0.0)
+#  amount_open               :decimal(, )
+#  discarded_at              :datetime
+#  issued_at                 :datetime
+#  locale                    :string
+#  payable_until             :datetime
+#  payment_info_type         :string
+#  payment_ref               :string
+#  payment_required          :boolean          default(TRUE)
+#  ref                       :string
+#  sent_at                   :datetime
+#  sequence_number           :integer
+#  sequence_year             :integer
+#  text                      :text
+#  type                      :string
+#  created_at                :datetime         not null
+#  updated_at                :datetime         not null
+#  booking_id                :uuid
+#  sent_with_notification_id :bigint
+#  supersede_invoice_id      :bigint
 #
 
 module Invoices