+++ spec/domain/booking_actions/manage/email_contract_spec.rb
@@ -2,11 +2,12 @@
 
 require 'rails_helper'
 
-describe BookingActions::Manage::EmailContract do
-  subject(:action) { described_class.new(booking) }
+describe BookingActions::EmailContract do
+  subject(:action) { described_class.new(booking, :email_contract) }
   subject(:invoke) { action.invoke }
   subject(:booking_after_invoke) do
     invoke
+    booking.notifications.each(&:deliver)
     booking
   end
 
@@ -15,15 +16,13 @@ describe BookingActions::Manage::EmailContract do
   let(:initial_state) { :definitive_request }
   let!(:contract) { create(:contract, booking:) }
 
-  describe '#allowed?' do
-    subject(:allowed) { action.allowed? }
-
-    it { expect(allowed).to be_truthy }
+  describe '#invokable?' do
+    it { expect(action.invokable?).to be_truthy }
   end
 
   describe '#invoke' do
     it { expect(invoke.success).to be_truthy }
-    it { expect(booking_after_invoke).to notify(:awaiting_contract_notification).to(:tenant) }
+    it { expect(booking_after_invoke).to notify(:email_contract_notification).to(:tenant) }
     it { expect(booking_after_invoke.contract).to be_sent }
     it { expect(booking_after_invoke.notifications.last.attachments).to be_present }
 
@@ -32,20 +31,19 @@ describe BookingActions::Manage::EmailContract do
 
       it do
         expect(deposit.invoice_parts).to be_present
-        invoke
-        expect(deposit.reload).to be_sent
+        expect(booking_after_invoke.invoices.find(deposit.id)).to be_sent
       end
     end
   end
 
-  describe '#label' do
-    subject(:label) { action.label }
+  describe '#invokable_with' do
+    subject(:invokable_with) { action.invokable_with }
 
-    it { expect(label).to eq(I18n.t('booking_actions.manage.email_contract.label_without_deposit')) }
+    it { expect(invokable_with[:label]).to eq(I18n.t('booking_actions.email_contract.label_without_deposit')) }
 
     context 'with deposit' do
       let!(:deposit) { Invoice::Factory.new(booking).build(type: Invoices::Deposit.to_s).tap(&:save) }
-      it { expect(label).to eq(I18n.t('booking_actions.manage.email_contract.label_with_deposit')) }
+      it { expect(invokable_with[:label]).to eq(I18n.t('booking_actions.email_contract.label_with_deposit')) }
     end
   end
 end